requirements.txt
aioconsole==0.5.1
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
iniconfig==1.1.1
logger==1.4
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
PyYAML==6.0
requests==2.28.1
tomli==2.0.1
urllib3==1.26.12
websockets==10.3
yarl==1.8.1
